requirements.txt
discord
python-dotenv
playwright